+++ CHANGES.rst
@@ -15,6 +15,7 @@ development
     - Fix access by honoring throttling employed by DPMA
     - Refactor and improve library API and inline documentation
     - Provide new access methods ``fetch_st36xml`` and ``fetch_pdf``
+    - Switch scraper from "mechanize" to "MechanicalSoup"
 
 - Upgrade to Moment.js 2.20.0 re. CVE-2016-4055